+++ src/Controller/UserController.php
@@ -8,6 +8,9 @@ use App\Repository\UserRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\CollectionType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
@@ -33,12 +36,14 @@ class UserController extends AbstractController
     {
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
+
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $encoded = $encoder->encodePassword($user, $user->getPassword());
             $user->setPassword($encoded);
+            $user->setActif('oui');
             $entityManager->persist($user);
             $entityManager->flush();
 
@@ -67,23 +72,50 @@ class UserController extends AbstractController
     public function edit(Request $request, User $user, UserPasswordEncoderInterface $encoder): Response
     {
         $form = $this->createForm(UserType::class, $user);
+
+        if ( $this->getUser()->getRoles()[0] == 'ROLE_ADMIN' ) {
+            $form->add('actif', ChoiceType::class, [
+                'choices'  => [
+                    'Oui' => 'oui',
+                    'Non' => 'non',
+                ]
+            ])
+            ->add('roles', CollectionType::class, [
+                'entry_type'   => ChoiceType::class,
+                'entry_options'  => [
+                    'choices'  => [
+                        'ROLE_USER' => 'ROLE_USER',
+                        'ROLE_ADMIN'     => 'ROLE_ADMIN',
+                    ],
+                ],
+            ]);
+        } 
+
+        $form->add('passwd', PasswordType::class, [
+            'required' => false,
+            'label' => 'Mot de passe'
+        ]);
+
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
-            $encoded = $encoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($encoded);
+            
+            if ( $form->get('passwd')->getData() != '' ) {
+                $encoded = $encoder->encodePassword($user, $user->getPasswd());
+                $user->setPassword($encoded);    
+            }
+            
             $entityManager->persist($user);
             $entityManager->flush();
 
             return $this->redirectToRoute('user_index', [
-                'id' => $user->getId(),
+              'id' => $user->getId(),
             ]);
         }
 
         if ( $user->getMail() == $this->getUser()->getMail() || $this->getUser()->getRoles()[0] == 'ROLE_ADMIN' )
         {
-            //var_dump($form->get('password')->getData());
             return $this->render('user/edit.html.twig', [
                 'user' => $user,
                 'form' => $form->createView(),
@@ -104,12 +136,13 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($user);
-            $entityManager->flush();
+        if ( $this->getUser()->getRoles()[0] == 'ROLE_ADMIN' ) {
+            if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($user);
+                $entityManager->flush();
+            }
         }
-
         return $this->redirectToRoute('user_index');
     }
 }
